Fix gate training: fp32 cache, gate=-2.0, zero W_out
Gate stuck at init because bfloat16 quant rounds tiny gradient updates
to zero at -4.0, and zero W_out kills gate gradient at step 0.
Keep cache float32 (6MB), gate=-2.0 (sigmoid=0.12).

+++ bam/train_cache.py
@@ -296,7 +296,7 @@ def main() -> None:
     device = next(model.parameters()).device
     cache = StateCache(
         d_model=hidden_size, d_attn=D_ATTN, max_entries=MAX_ENTRIES
-    ).to(device=device, dtype=torch.bfloat16)
+    ).to(device=device)  # float32: tiny module, gate needs fp32 precision
     cache.train()
     for p in cache.parameters():
         p.requires_grad = True
